feat: reorganize microsite components (John Correa)
