+++ src/components/demo/microsite/MicrositeOpenPlayground.jsx
@@ -1,16 +1,23 @@
 import * as React from "react";
 import { useRef, useState } from 'react';
-import { ColorPicker } from '@/components/HexColorPicker';
-import { List, arrayMove } from "react-movable";
-import { Pencil, Trash, Plus, Palette, ImagePlus } from "lucide-react";
-import { ImageToBase64 } from '@/components/PopupLogoMicrositeOpen';
-import { ContrastText } from '@/components/AutomaticTextContrast';
+import { ColorPicker } from '@/components/demo/microsite/HexColorPicker';
+import { ImageToBase64 } from '@/components/demo/microsite/PopupLogoMicrositeOpen';
+import { ContrastText } from '@/components/demo/microsite/AutomaticTextContrast';
+import { DragAndDrop, handleFieldSelection } from '@/components/demo/microsite/DragAndDrop';
+import { Palette, ImagePlus } from "lucide-react";
 import Image from 'next/image';
 import footerSvg from '@/images/logos/footer.svg';
 import generatePDF, { Margin } from 'react-to-pdf';
 
 export function MicrositeOpenPlayground() {
-  const [fields, setFields] = React.useState([
+  const targetRef = useRef();
+  const [isVisibleBtn, setIsVisibleBtn] = useState(false);
+  const [isVisibleLogo, setIsVisibleLogo] = useState(false);
+  const [isVisibleSidePanel, setIsVisibleSidePanel] = useState(false);
+  const [isVisiblePDF, setIsVisiblePDF] = useState(true);
+  const [colorBtn, setColorBtn] = useState("#64748B");
+  const [colorSidePanel, setColorSidePanel] = useState("#F1F5F9");
+  const [fields, setFields] = useState([
     { id: "reference", label: "Referencia", subLabel: "(Referencia)", type: "text", required: true },
     { id: "payment_description", label: "Descripción del pago", subLabel: "(Descripción del pago)", type: "text", required: true },
     { id: "currency", label: "Moneda", subLabel: "(Moneda)", type: "select", options: ["Selecciona una opción", "Peso colombiano", "Dólar estadounidense"], required: true },
@@ -21,63 +28,7 @@ export function MicrositeOpenPlayground() {
     { id: "buyer_surname", label: "Apellido del comprador", type: "text" },
     { id: "buyer_email", label: "Correo electrónico", type: "email" },
   ]);
-
-  const [fieldCounter, setFieldCounter] = React.useState(9); // Starting from 9 since initial items go up to 8
-  const protectedFields = ["reference", "payment_description", "currency", "amount"];
-
-  const handleEdit = (id) => {
-    const newLabel = prompt("Ingrese el nuevo nombre del campo:");
-    if (newLabel) {
-      setFields(fields.map(field => {
-        if (field.id === id) {
-          return { ...field, label: newLabel };
-        } else if (field.columns) {
-          return {
-            ...field,
-            columns: field.columns.map(col => col.id === id ? { ...col, label: newLabel } : col)
-          };
-        }
-        return field;
-      }));
-    }
-  };
-
-  const handleDelete = (id) => {
-    setFields(fields.map(field => {
-      if (field.id === id) {
-        return null;
-      } else if (field.columns) {
-        return {
-          ...field,
-          columns: field.columns.filter(col => col.id !== id)
-        };
-      }
-      return field;
-    }).filter(Boolean));
-  };
-
-  const addNewRow = (numFields) => {
-    const newFields = Array(numFields).fill(null).map((_, index) => ({
-      id: `field_${fieldCounter + index}`,
-      label: `Campo ${fieldCounter + index}`,
-      type: "text",
-    }));
-    setFieldCounter(fieldCounter + numFields);
-    setFields([...fields, { id: `row_${fieldCounter}`, columns: newFields }]);
-  };
-
-  const handleFieldSelection = (e) => {
-    const numFields = parseInt(e.target.value);
-    if (!isNaN(numFields)) addNewRow(numFields);
-  };
-
-  const targetRef = useRef();
-  const [isVisibleBtn, setIsVisibleBtn] = useState(false);
-  const [isVisibleLogo, setIsVisibleLogo] = useState(false);
-  const [isVisibleSidePanel, setIsVisibleSidePanel] = useState(false);
-  const [isVisiblePDF, setIsVisiblePDF] = useState(true);
-  const [colorBtn, setColorBtn] = useState("#64748B");
-  const [colorSidePanel, setColorSidePanel] = useState("#F1F5F9");
+  const [fieldCounter, setFieldCounter] = useState(9);
 
   const handleClickSidePanel = () => {
     setIsVisibleSidePanel(!isVisibleSidePanel);
@@ -124,6 +75,9 @@ export function MicrositeOpenPlayground() {
     }, 1000);
   };
 
+  const handleOnChange = (e) => {
+    handleFieldSelection(e, fields, setFields, fieldCounter, setFieldCounter);
+  };
 
   return (
     <div>
@@ -133,7 +87,7 @@ export function MicrositeOpenPlayground() {
           <select
             id="fieldSelector"
             className="p-2 bg-gray-500 text-white rounded-lg flex items-center justify-center cursor-pointer focus:outline-none hover:bg-gray-600"
-            onChange={handleFieldSelection}
+            onChange={handleOnChange}
           >
             <option value="">Agregar Campo</option>
             <option value="1">1 Fila de 1 Campo</option>
@@ -163,135 +117,12 @@ export function MicrositeOpenPlayground() {
                     <form noValidate="" className="p-0 lg:p-0 md:p-2 flex flex-col">
                       <div className="form-layout w-full">
 
-                        <List
-                          values={fields}
-                          onChange={({ oldIndex, newIndex }) =>
-                            setFields(arrayMove(fields, oldIndex, newIndex))
-                          }
-                          renderList={({ children, props, isDragged }) => (
-                            <div
-                              {...props}
-                              className="grid grid-cols-1 md:grid-cols-2 gap-4"
-                              style={{ cursor: isDragged ? "grabbing" : undefined }}
-                            >
-                              {children}
-                            </div>
-                          )}
-                          renderItem={({ value, props, isDragged, isSelected }) => {
-                            if (value.columns) {
-                              return (
-                                <div
-                                  {...props}
-                                  key={props.key}
-                                  style={{
-                                    ...props.style,
-                                    padding: "0.3rem",
-                                    cursor: isDragged ? "grabbing" : "grab",
-                                    borderRadius: "0.5rem",
-                                    backgroundColor: isDragged || isSelected ? "#EEE" : "#FFF",
-                                    zIndex: isDragged ? 1000 : "auto",
-                                    display: "grid",
-                                    gridTemplateColumns: `repeat(${value.columns.length}, 1fr)`,
-                                    gap: "8px",
-                                    width: "100%",
-                                  }}
-                                  className="col-span-2"
-                                >
-                                  {value.columns.map((field) => (
-                                    <div key={field.id} className="flex flex-col">
-                                      <div className="flex items-center gap-2">
-                                        <label className="text-sm font-medium text-gray-700 mb-2">{field.label}</label>
-                                      </div>
-                                      <div className="flex items-center gap-2">
-                                        {field.type === "select" ? (
-                                          <select
-                                            id={field.id}
-                                            className="w-full border border-gray-400 py-0 text-gray-900 outline-none focus:border-primary-300 rounded-lg h-10 flex items-center"
-                                          >
-                                            {field.options.map((option, index) => (
-                                              <option key={index} value={option}>
-                                                {option}
-                                              </option>
-                                            ))}
-                                          </select>
-                                        ) : (
-                                          <input
-                                            type="text"
-                                            className="w-full border border-gray-400 py-2 text-gray-900 outline-none placeholder:text-gray-400 focus:border-primary-300 rounded-lg"
-                                            id={field.id}
-                                          />
-                                        )}
-                                        {isVisiblePDF && (
-                                          <button type="button" onClick={() => handleEdit(field.id)} className="text-blue-500">
-                                            <Pencil size={16} />
-                                          </button>
-                                        )}
-                                        {isVisiblePDF && !protectedFields.includes(field.id) && (
-                                          <button type="button" onClick={() => handleDelete(field.id)} className="text-red-500">
-                                            <Trash size={16} />
-                                          </button>
-                                        )}
-                                      </div>
-                                    </div>
-                                  ))}
-                                </div>
-                              );
-                            } else {
-                              const isFullWidth =
-                                value.id === "reference" || value.id === "payment_description" || value.id === "buyer_email";
-                              return (
-                                <div
-                                  {...props}
-                                  key={props.key}
-                                  style={{
-                                    ...props.style,
-                                    padding: "0.3rem",
-                                    cursor: isDragged ? "grabbing" : "grab",
-                                    borderRadius: "0.5rem",
-                                    backgroundColor: isDragged || isSelected ? "#EEE" : "#FFF",
-                                    zIndex: isDragged ? 1000 : "auto",
-                                  }}
-                                  className={`flex flex-col ${isFullWidth ? "col-span-2" : "col-span-1"}`}
-                                >
-                                  <div className="flex items-center gap-2">
-                                    <label className="text-sm font-medium text-gray-700 mb-2">{value.label}</label>
-                                    <span className="text-sm text-gray-500 mb-2">{value.subLabel}</span>
-                                    {protectedFields.includes(value.id) && <p className="text-red-500">*</p>}
-                                  </div>
-                                  <div className="flex items-center gap-2">
-                                    {value.type === "select" ? (
-                                      <select
-                                        id={value.id}
-                                        className="w-full border border-gray-400 py-0 text-gray-900 outline-none focus:border-primary-300 rounded-lg h-10 flex items-center"
-                                      >
-                                        {value.options.map((option, index) => (
-                                          <option key={index} value={option}>
-                                            {option}
-                                          </option>
-                                        ))}
-                                      </select>
-                                    ) : (
-                                      <input
-                                        type="text"
-                                        className="w-full border border-gray-400 py-2 text-gray-900 outline-none placeholder:text-gray-400 focus:border-primary-300 rounded-lg"
-                                        id={value.id}
-                                      />
-                                    )}
-                                    {isVisiblePDF && (
-                                      <button type="button" onClick={() => handleEdit(value.id)} className="text-blue-500">
-                                        <Pencil size={16} />
-                                      </button>
-                                    )}
-                                    {isVisiblePDF && !protectedFields.includes(value.id) && (
-                                      <button type="button" onClick={() => handleDelete(value.id)} className="text-red-500">
-                                        <Trash size={16} />
-                                      </button>
-                                    )}
-                                  </div>
-                                </div>
-                              );
-                            }
-                          }}
+                        <DragAndDrop 
+                          isVisiblePDF={isVisiblePDF}
+                          fields={fields}
+                          setFields={setFields}
+                          fieldCounter={fieldCounter}
+                          setFieldCounter={setFieldCounter}
                         />
 
                         <div className="mt-6">
