+++ src/pages/payments/microsite-open/introduction.mdx
@@ -2,9 +2,9 @@ import { BenefitsOpen } from '@/components/BenefitsOpen';
 import { Requirements } from '@/components/Requirements';
 import { HeroPattern } from '@/components/HeroPattern';
 import { CenteredContainer } from '@/components/CenteredContainer';
-import { PopupComponent } from '@/components/PopupComponent';
-import { PopupContentMicrositeOpen } from '@/components/PopupContentMicrositeOpen';
-import { ContrastText } from '@/components/AutomaticTextContrast';
+import { PopupComponent } from '@/components/demo/microsite/PopupComponent';
+import { PopupContentMicrositeOpen } from '@/components/demo/microsite/PopupContentMicrositeOpen';
+import { ContrastText } from '@/components/demo/microsite/AutomaticTextContrast';
 
 
 export const sectionMode = 'nav';
